+++ service/controller/resource/tccpn/template/template_main_iam_policies.go
@@ -170,11 +170,11 @@ const TemplateMainIAMPolicies = `
           {{- if or (eq .IAMPolicies.Region "cn-north-1") (eq .IAMPolicies.Region "cn-northwest-1") }}
           - Effect: "Allow"
             Principal:
-              Federated: "arn:{{ .IAMPolicies.RegionARN }}:iam::{{ .IAMPolicies.AccountID }}:oidc-provider/s3.{{ .IAMPolicies.Region }}.amazonaws.com.cn/{{ .IAMPolicies.AccountID }}-g8s-{{ .IAMPolicies.ClusterID }}-oidc-pod-identity"
+              Federated: "arn:{{ .IAMPolicies.RegionARN }}:iam::{{ .IAMPolicies.AccountID }}:oidc-provider/s3.{{ .IAMPolicies.Region }}.amazonaws.com.cn/{{ .IAMPolicies.AccountID }}-g8s-{{ .IAMPolicies.ClusterID }}-oidc-pod-identity-v2"
             Action: "sts:AssumeRoleWithWebIdentity"
             Condition:
               StringLike:
-                "s3.{{ .IAMPolicies.Region }}.amazonaws.com.cn/{{ .IAMPolicies.AccountID }}-g8s-{{ .IAMPolicies.ClusterID }}-oidc-pod-identity:sub": "system:serviceaccount:*:aws-load-balancer-controller*"
+                "s3.{{ .IAMPolicies.Region }}.amazonaws.com.cn/{{ .IAMPolicies.AccountID }}-g8s-{{ .IAMPolicies.ClusterID }}-oidc-pod-identity-v2:sub": "system:serviceaccount:*:aws-load-balancer-controller*"
           {{- end }}
           {{- if ne .IAMPolicies.CloudfrontDomain "" }}
           - Effect: "Allow"
@@ -384,7 +384,7 @@ const TemplateMainIAMPolicies = `
             Action: "sts:AssumeRoleWithWebIdentity"
             Condition:
               StringLike:
-                "{{ .IAMPolicies.CloudfrontDomain }}:sub": "system:serviceaccount:*:external-dns*"
+                "{{ .IAMPolicies.CloudfrontDomain }}:sub": "system:serviceaccount:*:*external-dns*"
           {{- end }}
           {{- if ne .IAMPolicies.CloudfrontAliasDomain "" }}
           - Effect: "Allow"
@@ -393,7 +393,7 @@ const TemplateMainIAMPolicies = `
             Action: "sts:AssumeRoleWithWebIdentity"
             Condition:
               StringLike:
-                "{{ .IAMPolicies.CloudfrontAliasDomain }}:sub": "system:serviceaccount:*:external-dns*"
+                "{{ .IAMPolicies.CloudfrontAliasDomain }}:sub": "system:serviceaccount:*:*external-dns*"
           {{- end }}
   Route53ManagerRolePolicy:
     Type: "AWS::IAM::Policy"