+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Check if all nodes are rolled before deleting AWS CNI resources when upgrading from v18 to v19.
+- Change Route53 Trust Policy to allow other applications to use the role.
+
+### Fix
+- Change AWS LB Controller Trust Policy for the new S3 bucket.
 
 ## [14.17.1] - 2023-05-11