+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for customizing `controller-manager` `terminated-pod-gc-threshold` value through annotation `controllermanager.giantswarm.io/terminated-pod-gc-threshold`
+- Add Service Monitor.
 
 ### Changed
 
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change Route53 Trust Policy to allow other applications to use the role.
 
 ### Fix
+
 - Change AWS LB Controller Trust Policy for the new S3 bucket.
 
 ## [14.17.1] - 2023-05-11