+++ CHANGELOG.md
@@ -17,9 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Check if all nodes are rolled before deleting AWS CNI resources when upgrading from v18 to v19.
 - Change Route53 Trust Policy to allow other applications to use the role.
 
-### Fix
+### Fixed
 
 - Change AWS LB Controller Trust Policy for the new S3 bucket.
+- Added pss exceptions for volumes.
 
 ## [14.17.1] - 2023-05-11